% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkNebulaGpuParams}
\alias{checkNebulaGpuParams}
\title{Check GPU NEBULA params}
\usage{
checkNebulaGpuParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of \code{\link[=params_nebula_gpu]{params_nebula_gpu()}}.
}
\keyword{internal}
