% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkNnParamsGpu}
\alias{checkNnParamsGpu}
\title{Check GPU nearest neighbour parameters}
\usage{
checkNnParamsGpu(x)
}
\arguments{
\item{x}{The list to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the GPU nearest neighbour
parameters.
}
\keyword{internal}
