% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkScFastmnnGpu}
\alias{checkScFastmnnGpu}
\title{Check GPU fastMNN parameters}
\usage{
checkScFastmnnGpu(x)
}
\arguments{
\item{x}{The list to check/assert.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the GPU fastMNN parameters.
}
\keyword{internal}
