% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R, R/param_wrappers.R
\name{checkScHarmonyParamsV2Gpu}
\alias{checkScHarmonyParamsV2Gpu}
\title{Check Harmony v2 GPU parameters}
\usage{
checkScHarmonyParamsV2Gpu(x)

checkScHarmonyParamsV2Gpu(x)
}
\arguments{
\item{x}{The list to check/assert.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.

\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking Harmony v2 GPU parameters.

Checkmate extension for checking Harmony v2 GPU parameters.
}
\keyword{internal}
