% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkScHarmonyV2GpuParams}
\alias{checkScHarmonyV2GpuParams}
\title{Check Harmony v2 GPU params}
\usage{
checkScHarmonyV2GpuParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of
\code{\link[=params_sc_harmony_v2_gpu]{params_sc_harmony_v2_gpu()}}.
}
\keyword{internal}
