% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkScIvfParams}
\alias{checkScIvfParams}
\title{Check IVF GPU parameters}
\usage{
checkScIvfParams(x)
}
\arguments{
\item{x}{The list to check/assert.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking IVF GPU parameters.
}
\keyword{internal}
