% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkTsneGpuParams}
\alias{checkTsneGpuParams}
\title{Check GPU t-SNE params}
\usage{
checkTsneGpuParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of \code{\link[=params_tsne_gpu]{params_tsne_gpu()}}.
}
\keyword{internal}
