% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkTsneParamsGpu}
\alias{checkTsneParamsGpu}
\title{Check t-SNE parameters (GPU version)}
\usage{
checkTsneParamsGpu(x)
}
\arguments{
\item{x}{The list to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the t-SNE parameters (GPU
version).
}
\keyword{internal}
