% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkUmapGpuParams}
\alias{checkUmapGpuParams}
\title{Check GPU UMAP params}
\usage{
checkUmapGpuParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of \code{\link[=params_umap_gpu]{params_umap_gpu()}}.
}
\keyword{internal}
