% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkUmapParamsGpu}
\alias{checkUmapParamsGpu}
\title{Check UMAP parameters (GPU version)}
\usage{
checkUmapParamsGpu(x)
}
\arguments{
\item{x}{The list to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the UMAP parameters (GPU
version).
}
\keyword{internal}
