% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_processing.R
\name{.set_snn_graph_gpu}
\alias{.set_snn_graph_gpu}
\title{Build an sNN igraph from kNN data and attach it to a single cell object}
\usage{
.set_snn_graph_gpu(object, knn_data, snn_params, modality, .verbose)
}
\arguments{
\item{object}{\code{SingleCells} (or \code{SingleCellsMultiModal}) class.}

\item{knn_data}{Initialised \code{sc_knn} with the kNN data.}

\item{snn_params}{List. Output of \code{\link[bixverse:params_sc_neighbours]{bixverse::params_sc_neighbours()}}.}

\item{modality}{String. One of \code{c("rna", "adt")}.}

\item{.verbose}{Boolean or integer. Controls verbosity.}
}
\value{
The object with the sNN graph in the selected modality slot.
}
\description{
Build an sNN igraph from kNN data and attach it to a single cell object
}
\keyword{internal}
