% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{membership}
\alias{membership}
\alias{membership.KMeansClusterGPU}
\title{Get cluster assignments}
\usage{
membership(x)

\method{membership}{KMeansClusterGPU}(x)
}
\arguments{
\item{x}{A \code{KMeansClusterGPU} object.}
}
\value{
Integer vector of length samples with cluster assignments
(1-indexed).
}
\description{
Get cluster assignments
}
