% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_kmeans_cluster_gpu}
\alias{new_kmeans_cluster_gpu}
\title{Construct a KMeansClusterGPU object}
\usage{
new_kmeans_cluster_gpu(centroids, assignments, k, metric)
}
\arguments{
\item{centroids}{Numeric matrix of shape k x features.}

\item{assignments}{Integer vector of length samples (1-indexed).}

\item{k}{Integer. Number of clusters.}

\item{metric}{Character. Distance metric used.}
}
\value{
A \code{KMeansClusterGPU} S3 object.
}
\description{
Construct a KMeansClusterGPU object
}
\keyword{internal}
