% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_kmeans_gpu}
\alias{params_kmeans_gpu}
\title{Default parameters for GPU k-means}
\usage{
params_kmeans_gpu(
  k_means_iter = 50L,
  k_means_init = NULL,
  metric = c("euclidean", "cosine"),
  fixed = FALSE,
  quantise = FALSE
)
}
\arguments{
\item{k_means_iter}{Integer. Number of k-means iterations. Defaults to \code{50L}.}

\item{k_means_init}{String or \code{NULL}. Initialisation method. One of
\code{"random"}, \code{"parallel"}, or \code{"plusplus"}. If \code{NULL}, determined on the Rust
side. Defaults to \code{NULL}.}

\item{metric}{String. The distance metric. One of \code{c("euclidean", "cosine")}.
Defaults to \code{"euclidean"}.}

\item{fixed}{Boolean. Shall the algorithm be run for a fixed number of
iterations, without checking for convergence. Defaults to \code{FALSE}.}

\item{quantise}{Boolean. Whether to quantise data to \code{fp16} before
clustering. This can improve performance in circumstances where it is memory
bound. Defaults to \code{FALSE}.}
}
\value{
A named list with the following elements:
\itemize{
\item k_means_iter - Integer. Number of k-means iterations. Defaults to
\code{50L}.
\item k_means_init - String or \code{NULL}. Initialisation method. One of
\code{"random"}, \code{"parallel"}, or \code{"plusplus"}. If \code{NULL}, determined on the Rust
side. Defaults to \code{NULL}.
\item metric - String. The distance metric. One of \code{c("euclidean", "cosine")}. Defaults to \code{"euclidean"}.
\item fixed - Boolean. Shall the algorithm be run for a fixed number of
iterations, without checking for convergence. Defaults to \code{FALSE}.
\item quantise - Boolean. Whether to quantise data to \code{fp16} before
clustering. This can improve performance in circumstances where it is memory
bound. Defaults to \code{FALSE}.
}
}
\description{
Default parameters for GPU k-means
}
