% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_parametric_umap}
\alias{params_parametric_umap}
\title{Wrapper function to generate parametric UMAP parameters}
\usage{
params_parametric_umap(
  local_connectivity = 1,
  bandwidth = 1e-05,
  mix_weight = 1,
  hidden_layers = c(128L, 64L, 32L),
  lr = 0.001,
  corr_weight = 0,
  n_epochs = 500L,
  batch_size = 256L,
  neg_sample_rate = 5L
)
}
\arguments{
\item{local_connectivity}{Numeric. Number of nearest neighbours assumed to be
at distance zero. Defaults to \code{1.0}.}

\item{bandwidth}{Numeric. Convergence tolerance for smooth kNN distance
binary search. Defaults to \code{1e-05}.}

\item{mix_weight}{Numeric. Balance between fuzzy union and directed graph
during symmetrisation. Defaults to \code{1.0}.}

\item{hidden_layers}{Integer vector. Hidden layer sizes for the MLP encoder.
Defaults to \code{c(128L, 64L, 32L)}.}

\item{lr}{Numeric. Learning rate for the neural network optimiser. Defaults
to \code{0.001}.}

\item{corr_weight}{Numeric. Coefficient for the negative Pearson correlation
loss that encourages similar distances in embedding and original space.
Defaults to \code{0.0}.}

\item{n_epochs}{Integer. Number of training epochs. Defaults to \code{500L}.}

\item{batch_size}{Integer. Training batch size. Defaults to \code{256L}.}

\item{neg_sample_rate}{Integer. Number of negative samples per positive edge.
Defaults to \code{5L}.}
}
\value{
A named list with the following elements:
\itemize{
\item local_connectivity - Numeric. Number of nearest neighbours assumed to
be at distance zero. Defaults to \code{1.0}.
\item bandwidth - Numeric. Convergence tolerance for smooth kNN distance
binary search. Defaults to \code{1e-05}.
\item mix_weight - Numeric. Balance between fuzzy union and directed graph
during symmetrisation. Defaults to \code{1.0}.
\item hidden_layers - Integer vector. Hidden layer sizes for the MLP
encoder. Defaults to \code{c(128L, 64L, 32L)}.
\item lr - Numeric. Learning rate for the neural network optimiser. Defaults
to \code{0.001}.
\item corr_weight - Numeric. Coefficient for the negative Pearson
correlation loss that encourages similar distances in embedding and original
space. Defaults to \code{0.0}.
\item n_epochs - Integer. Number of training epochs. Defaults to \code{500L}.
\item batch_size - Integer. Training batch size. Defaults to \code{256L}.
\item neg_sample_rate - Integer. Number of negative samples per positive
edge. Defaults to \code{5L}.
}
}
\description{
Wrapper function to generate parametric UMAP parameters
}
