% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_sc_ivf}
\alias{params_sc_ivf}
\title{Default parameters for IVF-GPU kNN search}
\usage{
params_sc_ivf(
  k = 15L,
  ann_dist = "euclidean",
  nlist = NULL,
  nprobe = NULL,
  nquery = NULL,
  max_iters = NULL,
  seed = 42L
)
}
\arguments{
\item{k}{Integer. Number of neighbours to identify.}

\item{ann_dist}{Character. Distance metric to use. One of \code{"euclidean"} or
\code{"cosine"}.}

\item{nlist}{Optional integer. Number of clusters to partition the index
into. If \code{NULL}, defaults to \code{sqrt(n)}.}

\item{nprobe}{Optional integer. Number of clusters to probe at query time.
If \code{NULL}, defaults to \code{sqrt(nlist)}.}

\item{nquery}{Optional integer. Number of query vectors processed per GPU
batch. If \code{NULL}, defaults to 100,000.}

\item{max_iters}{Optional integer. Maximum k-means iterations during index
build. If \code{NULL}, defaults to 30.}

\item{seed}{Integer. Seed for k-means initialisation.}
}
\value{
A list with the parameters.
}
\description{
Default parameters for IVF-GPU kNN search
}
