% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_umap_gpu}
\alias{params_umap_gpu}
\title{Wrapper function to generate UMAP parameters (GPU version)}
\usage{
params_umap_gpu(
  local_connectivity = 1,
  bandwidth = 1e-05,
  mix_weight = 1,
  lr = 1,
  n_epochs = NULL,
  neg_sample_rate = 5L,
  gamma = 1,
  optimiser = c("adam_gpu", "adam_parallel", "sgd", "adam"),
  init = c("spectral", "pca", "random"),
  randomised = FALSE
)
}
\arguments{
\item{local_connectivity}{Numeric. Number of nearest neighbours assumed to be
at distance zero. Defaults to \code{1.0}.}

\item{bandwidth}{Numeric. Convergence tolerance for smooth kNN distance
binary search. Defaults to \code{1e-05}.}

\item{mix_weight}{Numeric. Balance between fuzzy union and directed graph
during symmetrisation. Defaults to \code{1.0}.}

\item{lr}{Numeric. Learning rate. Defaults to \code{1.0}.}

\item{n_epochs}{Integer or \code{NULL}. Number of optimisation epochs. Defaults to
\code{NULL}.}

\item{neg_sample_rate}{Integer. Number of negative samples per positive
sample. Defaults to \code{5L}.}

\item{gamma}{Numeric. Repulsion strength. Defaults to \code{1.0}.}

\item{optimiser}{String. The optimiser for the embedding. One of
\code{c("adam_gpu", "adam_parallel", "sgd", "adam")}. Defaults to \code{"adam_gpu"}.}

\item{init}{String. Embedding initialisation method. One of \code{c("spectral", "pca", "random")}. Defaults to \code{"spectral"}.}

\item{randomised}{Boolean. Use randomised SVD for PCA initialisation.
Defaults to \code{FALSE}.}
}
\value{
A named list with the following elements:
\itemize{
\item local_connectivity - Numeric. Number of nearest neighbours assumed to
be at distance zero. Defaults to \code{1.0}.
\item bandwidth - Numeric. Convergence tolerance for smooth kNN distance
binary search. Defaults to \code{1e-05}.
\item mix_weight - Numeric. Balance between fuzzy union and directed graph
during symmetrisation. Defaults to \code{1.0}.
\item lr - Numeric. Learning rate. Defaults to \code{1.0}.
\item n_epochs - Integer or \code{NULL}. Number of optimisation epochs. Defaults
to \code{NULL}.
\item neg_sample_rate - Integer. Number of negative samples per positive
sample. Defaults to \code{5L}.
\item gamma - Numeric. Repulsion strength. Defaults to \code{1.0}.
\item optimiser - String. The optimiser for the embedding. One of
\code{c("adam_gpu", "adam_parallel", "sgd", "adam")}. Defaults to \code{"adam_gpu"}.
\item init - String. Embedding initialisation method. One of \code{c("spectral", "pca", "random")}. Defaults to \code{"spectral"}.
\item randomised - Boolean. Use randomised SVD for PCA initialisation.
Defaults to \code{FALSE}.
}
}
\description{
Wrapper function to generate UMAP parameters (GPU version)
}
