% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_verbosity}
\alias{parse_verbosity}
\title{Helper to parse the verbosity}
\usage{
parse_verbosity(input)
}
\arguments{
\item{input}{Boolean or integer to parse}
}
\value{
The integer controlling the verbosity
}
\description{
Helper to parse the verbosity
}
\keyword{internal}
