% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{predict.ParametricUmapModel}
\alias{predict.ParametricUmapModel}
\title{Predict embeddings for new data using a trained parametric UMAP model}
\usage{
\method{predict}{ParametricUmapModel}(object, newdata, ...)
}
\arguments{
\item{object}{A \code{ParametricUmapModel} object.}

\item{newdata}{Numerical matrix or data frame. New data of shape
samples x features. Must have the same number of features as the
training data.}

\item{...}{Ignored.}
}
\value{
A numerical matrix with dimensions samples x n_dim.
}
\description{
Predict embeddings for new data using a trained parametric UMAP model
}
