% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.ParametricUmapModel}
\alias{print.ParametricUmapModel}
\title{Print a parametric UMAP model}
\usage{
\method{print}{ParametricUmapModel}(x, ...)
}
\arguments{
\item{x}{A \code{ParametricUmapModel} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a parametric UMAP model
}
\keyword{internal}
