% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_cor_gpu}
\alias{rs_cor_gpu}
\title{GPU-accelerated correlation calculations}
\usage{
rs_cor_gpu(x, spearman, verbose)
}
\arguments{
\item{x}{Numerical matrix. The matrix for which to calculate the column
pairwise correlation matrix.}

\item{spearman}{Boolean. Shall the Spearman correlation be calculated
instead of Pearson.}

\item{verbose}{Boolean. Controls verbosity of the function.}
}
\value{
The correlation matrix
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
GPU-accelerated pairwise column correlations. Has the options of Pearson and
Spearman correlation coefficient calculations.
}
