% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_exhaustive_gpu_knn}
\alias{rs_exhaustive_gpu_knn}
\title{Generate an GPU-accelerated kNN graph from an exhaustive search}
\usage{
rs_exhaustive_gpu_knn(embd, k, dist_metric, verbose)
}
\arguments{
\item{embd}{Numeric matrix of embeddings, cells x features.}

\item{k}{Integer. Number of neighbours to return.}

\item{dist_metric}{String. Distance metric; one of
\code{c("euclidean", "cosine")}.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A named list with:
\itemize{
\item \code{indices} - Integer matrix of shape cells x k with 0-based neighbour
indices.
\item \code{dist} - Numeric matrix of shape cells x k with distances to the
neighbours.
\item \code{dist_metric} - Character. The distance metric used.
}
}
\description{
Runs an exhaustive kNN search on the GPU.
}
