% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_serialise_parametric_umap}
\alias{rs_serialise_parametric_umap}
\title{Takes in a parametric UMAP and serialises it to raw bytes}
\usage{
rs_serialise_parametric_umap(model)
}
\arguments{
\item{model}{Robject. The trained parametric UMAP model to serialise.}
}
\value{
The raw bytes of the model
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Serialises a trained parametric UMAP model to bytes for saving the data.
}
\keyword{internal}
