% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_embeddings.R
\name{save_parametric_umap}
\alias{save_parametric_umap}
\title{Save a parametric UMAP as a qs2 file}
\usage{
save_parametric_umap(model, path)
}
\arguments{
\item{model}{\code{ParametricUmapModel} you want to save.}

\item{path}{String. The path to the file. Needs to end with \code{".qs"} file
extension.}
}
\value{
Saves the file to the provided path.
}
\description{
Save a parametric UMAP as a qs2 file
}
