% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_gpu.R
\name{calculate_pca_gpu_sc}
\alias{calculate_pca_gpu_sc}
\title{GPU-accelerated PCA for single cell}
\usage{
calculate_pca_gpu_sc(object, no_pcs, hvg = NULL, seed = 42L, .verbose = TRUE)
}
\arguments{
\item{object}{\code{SingleCells} class}

\item{no_pcs}{Integer. Number of PCs to calculate.}

\item{hvg}{Optional integer. If you want to provide your own HVG genes.
Otherwise, the function will default to what is found in
\code{\link[bixverse:get_hvg]{bixverse::get_hvg()}}. Please provide 1-indexed genes here! If you provide
these, the internal HVG will be overwritten.}

\item{seed}{Integer. Controls reproducibility. Only relevant if
\code{randomised_svd = TRUE}.}

\item{.verbose}{Boolean or integer. Controls verbosity and returns run times.
\code{FALSE} -> quiet, \code{TRUE} or \code{1L} -> normal verbosity, \code{2L} -> detailed
verbosity.}
}
\value{
The function will add the PCA factors, loadings and singular values
to the object cache in memory.
}
\description{
This function will run sparse, randomised SVD while running several of the
large matrix multiplications on GPU for improved speed. This also means you
will have to provide the necessary VRAM for your data set. This version only
works on the \code{"rna"} modality.
}
