% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkParametricUmapParams}
\alias{checkParametricUmapParams}
\title{Check parametric UMAP parameters}
\usage{
checkParametricUmapParams(x)
}
\arguments{
\item{x}{The list to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the parametric UMAP
parameters.
}
\keyword{internal}
