% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_mnn_gpu.R
\name{.fast_mnn_gpu}
\alias{.fast_mnn_gpu}
\title{Shared implementation of the GPU fastMNN}
\usage{
.fast_mnn_gpu(
  object,
  batch_column,
  no_embd_to_use,
  fastmnn_params,
  seed,
  .verbose
)
}
\arguments{
\item{object}{\code{SingleCells} or \code{SingleCellsSubset} class from \code{bixverse}.}

\item{batch_column}{String. The column with the batch information in the
obs data of the class.}

\item{no_embd_to_use}{Optional integer. Number of PCs to use. If \code{NULL} all
will be used.}

\item{fastmnn_params}{List. Output of \code{\link[=params_sc_fastmnn_gpu]{params_sc_fastmnn_gpu()}}.}

\item{seed}{Integer. Random seed.}

\item{.verbose}{Boolean or integer. Controls verbosity and returns run
times. \code{FALSE} -> quiet, \code{TRUE} or \code{1L} -> normal verbosity, \code{2L} ->
detailed verbosity.}
}
\value{
The object with the \code{"mnn_gpu"} embedding set.
}
\description{
Body behind both \code{fast_mnn_gpu_sc()} methods.
}
\keyword{internal}
