% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_umap.R
\name{.prepare_parametric_umap_params}
\alias{.prepare_parametric_umap_params}
\title{Internal helper to prepare parametric UMAP parameters}
\usage{
.prepare_parametric_umap_params(
  min_dist,
  spread,
  knn_method,
  nn_params,
  parametric_umap_params
)
}
\arguments{
\item{min_dist}{Numeric. Minimum distance between embedded points.}

\item{spread}{Numeric. Effective scale of embedded points.}

\item{knn_method}{String. Approximate nearest neighbour method.}

\item{nn_params}{Named list. Nearest neighbour parameters, see
\code{\link[manifoldsR:params_nn]{manifoldsR::params_nn()}}.}

\item{parametric_umap_params}{Named list. Parametric UMAP parameters.}
}
\value{
Returns the merged list of final parameters.
}
\description{
Internal helper to prepare parametric UMAP parameters
}
\keyword{internal}
