% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne_gpu.R
\name{.prepare_tsne_params_gpu}
\alias{.prepare_tsne_params_gpu}
\title{Internal helper to prepare the t-SNE parameters (GPU version)}
\usage{
.prepare_tsne_params_gpu(knn_method, nn_params, tsne_params)
}
\arguments{
\item{knn_method}{String. Method to use to generate the kNN graph.}

\item{nn_params}{Named list. The nearest neighbour search parameters (GPU).}

\item{tsne_params}{Named list. The t-SNE-specific parameters (GPU).}
}
\value{
Returns the list of final parameters.
}
\description{
Internal helper to prepare the t-SNE parameters (GPU version)
}
\keyword{internal}
