% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_gpu.R
\name{.prepare_umap_params_gpu}
\alias{.prepare_umap_params_gpu}
\title{Internal helper to prepare the UMAP parameters (GPU version)}
\usage{
.prepare_umap_params_gpu(
  n,
  min_dist,
  spread,
  knn_method,
  nn_params,
  umap_params,
  .verbose = TRUE
)
}
\arguments{
\item{n}{Integer. Number of samples in the data set}

\item{min_dist}{Numeric. Minimum distance between embedded points.}

\item{spread}{Numeric. Effective scale of embedded points.}

\item{knn_method}{String. Method to use to generate the kNN graph.}

\item{nn_params}{Named list. The nearest neighbour search parameters (GPU).}

\item{umap_params}{Named list. The UMAP-specific parameters (GPU).}

\item{.verbose}{Boolean. Controls verbosity}
}
\value{
Returns the list of final parameters.
}
\description{
Internal helper to prepare the UMAP parameters (GPU version)
}
\keyword{internal}
