% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_gpu.R
\name{generate_knn_graph_gpu}
\alias{generate_knn_graph_gpu}
\title{Generate a k-nearest neighbour graph (GPU-accelerated)}
\usage{
generate_knn_graph_gpu(
  data,
  k,
  knn_method = c("nndescent", "exhaustive", "ivf"),
  nn_params = params_nn_gpu(),
  seed = 42L,
  extract_knn = lifecycle::deprecated(),
  .verbose = TRUE
)
}
\arguments{
\item{data}{Numeric matrix. The embedding or feature matrix to compute
neighbours on. Rows are observations, columns are features.}

\item{k}{Integer. The number of nearest neighbours to compute.}

\item{knn_method}{Character. The algorithm to use for nearest neighbour
search. One of \code{c("exhaustive", "ivf", "nndescent")}. Defaults to
\code{"nndescent"}}

\item{nn_params}{List. Output of \code{\link[=params_nn_gpu]{params_nn_gpu()}}.}

\item{seed}{Integer. For reproducibility. Defaults to \code{42L}.}

\item{extract_knn}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use the \code{extract_knn}
field of \code{\link[=params_nn_gpu]{params_nn_gpu()}} instead.}

\item{.verbose}{Boolean. Controls verbosity.}
}
\value{
A nearest neighbours class object with 1-indexed neighbour indices
and distances. Euclidean distances are true L2, not squared.
}
\description{
This function generates a kNN graph based on a given numeric matrix. Three
different GPU-accelerated versions are available
\itemize{
\item \code{"exhaustive"} - Exact nearest neighbour search via GPU.
\item \code{"ivf"} - Inverted file index that leverages k-means clustering
and probing a few of the clusters via GPU-accelerated distance
calculations.
\item \code{"nndescent"} - A CAGRA style nearest neighbour search on the GPU.
}
}
