% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_gpu.R
\name{k_means_cluster_gpu}
\alias{k_means_cluster_gpu}
\title{GPU-accelerated k-means clustering}
\usage{
k_means_cluster_gpu(
  data,
  k,
  kmeans_params = params_kmeans_gpu(),
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{data}{Numerical matrix or data frame. The data to cluster, of shape
samples x features. Will be coerced to a matrix.}

\item{k}{Integer. Number of clusters to create. Must be >= 2.}

\item{kmeans_params}{Named list. GPU-accelerated k-mean parameters, see
\code{\link[=params_kmeans_gpu]{params_kmeans_gpu()}}.}

\item{seed}{Integer. Random seed for reproducibility. Defaults to 42L.}

\item{.verbose}{Logical. Controls verbosity. Defaults to \code{TRUE}.}
}
\description{
Performs k-means clustering as a GPU-accelerated form. In its current form
uses the cubecl with wgpu backend.
}
