% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_umap.R
\name{load_parametric_umap}
\alias{load_parametric_umap}
\title{Load a parametric UMAP as a qs2 file}
\usage{
load_parametric_umap(path)
}
\arguments{
\item{path}{String. The path to the serialised trained parametric UMAP model
on disk.}
}
\value{
The \code{ParametricUmapModel}.
}
\description{
Load a parametric UMAP as a qs2 file
}
