% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_gpu.R
\name{parametric_umap}
\alias{parametric_umap}
\title{Parametric UMAP}
\usage{
parametric_umap(
  data,
  n_dim = 2L,
  k = 15L,
  min_dist = 0.1,
  spread = 1,
  knn_method = c("hnsw", "annoy", "nndescent", "balltree", "exhaustive"),
  nn_params = manifoldsR::params_nn(),
  parametric_umap_params = params_parametric_umap(),
  use_gpu = TRUE,
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{data}{Numerical matrix or data frame. The data to embed of shape
samples x features. Will be coerced to a matrix.}

\item{n_dim}{Integer. Number of embedding dimensions. Defaults to \code{2L}.}

\item{k}{Integer. Number of nearest neighbours. Defaults to \code{15L}.}

\item{min_dist}{Numeric. Minimum distance between embedded points. Defaults
to \code{0.1}.}

\item{spread}{Numeric. Effective scale of embedded points. Defaults to
\code{1.0}.}

\item{knn_method}{Character. Approximate nearest neighbour algorithm. One of
\code{"hnsw"}, \code{"annoy"}, \code{"nndescent"}, \code{"balltree"}, or \code{"exhaustive"}.
Defaults to \code{"hnsw"}.}

\item{nn_params}{Named list. Nearest neighbour parameters, see
\code{\link[manifoldsR:params_nn]{manifoldsR::params_nn()}}.}

\item{parametric_umap_params}{Named list. Parametric UMAP parameters, see
\code{\link[=params_parametric_umap]{params_parametric_umap()}}.}

\item{use_gpu}{Boolean. Shall the neural net be trained on GPU via the
\code{wgpu} backend. On smaller datasets, the CPU can be faster (via the
\code{ndarray}) backend due to kernel launch overhead.
data sets, the CPU will be faster via the Ndarray.}

\item{seed}{Integer. Random seed for reproducibility. Defaults to \code{42L}.}

\item{.verbose}{Logical. Controls verbosity. Defaults to \code{TRUE}.}
}
\value{
A \code{ParametricUmapModel} object containing the embedding matrix
and the trained encoder model.
}
\description{
Performs parametric UMAP dimensionality reduction using a
neural network encoder trained on the GPU via wgpu.
}
