% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_kmeans_gpu}
\alias{params_kmeans_gpu}
\title{Default parameters for GPU k-means}
\usage{
params_kmeans_gpu(
  k_means_iter = 50L,
  k_means_init = NULL,
  metric = c("euclidean", "cosine"),
  fixed = FALSE,
  quantise = FALSE
)
}
\arguments{
\item{k_means_iter}{Integer. Number of k-means iterations.}

\item{k_means_init}{Optional character. Initialisation method. One of
\code{"random"}, \code{"parallel"}, or \code{"plusplus"}. If \code{NULL}, determined on the
Rust side.}

\item{metric}{String. One of \code{c("euclidean", "cosine")}.}

\item{fixed}{Logical. Shall the algorithm be run for a fixed number of
iterations, without checking for convergence.}

\item{quantise}{Logical. Whether to quantise data to \code{fp16} before
clustering. This can improve performance in circumstances where it is
memory bound.}
}
\value{
A list with the parameters.
}
\description{
Default parameters for GPU k-means
}
