% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_nn_gpu}
\alias{params_nn_gpu}
\title{Wrapper function to generate GPU nearest neighbour parameters}
\usage{
params_nn_gpu(
  dist_metric = c("euclidean", "cosine"),
  n_list = NULL,
  n_probes = NULL,
  node_degree_final = NULL,
  k_build = NULL,
  n_tree = NULL,
  refine_sweeps = 0L,
  delta = 0.001,
  rho = NULL,
  beam_width = NULL,
  max_beam_iters = NULL,
  n_entry_points = NULL
)
}
\arguments{
\item{dist_metric}{Character. The distance metric to use. Defaults to
\code{"euclidean"}.}

\item{n_list}{Optional integer. IVF GPU: Number of clusters to use. If
\code{NULL}, will default to \code{sqrt(n)}.}

\item{n_probes}{Optional integer. IVF GPU: Number of clusters to probe. If
\code{NULL}, will default to \code{sqrt(n_list)}.}

\item{node_degree_final}{Optional integer. Final node degree of the CAGRA
navigational graph. If \code{NULL}, defaults to \code{30} on the Rust side.}

\item{k_build}{Optional integer. Number of k-neighbours during the
NNDescent build phase before CAGRA pruning. If \code{NULL}, defaults to
\code{1.5 * node_degree_final} on the Rust side. (Cannot be smaller than
\code{node_degree_final})}

\item{n_tree}{Optional integer. CAGRA GPU: Number of trees for graph build.
Automatically if \code{NULL}.}

\item{refine_sweeps}{Integer. Number of refinement sweeps during graph
generation.}

\item{delta}{Float. CAGRA GPU: Early termination parameter for NN descent.
Defaults to \code{0.001}.}

\item{rho}{Optional float. CAGRA GPU: Sample rate parameter for NN descent.
Defaults to \code{0.5} if not provided.}

\item{beam_width}{Optional integer. CAGRA GPU: Beam width for beam search. If
not provided will be set to \code{max(c(k, node_degree_final, 16L)) * 2}.}

\item{max_beam_iters}{Optional integer. CAGRA GPU: Maximum number of beam
search iterations. If not provided, defaults to \code{3 * beam_width}.}

\item{n_entry_points}{Optional integer. CAGRA GPU: Number of entry points for
beam search. If not provided, defaults to \code{8L}.}
}
\value{
A list with the GPU nearest neighbour parameters.
}
\description{
Wrapper function to generate GPU nearest neighbour parameters
}
