% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_nn_gpu}
\alias{params_nn_gpu}
\title{Wrapper function to generate GPU nearest neighbour parameters}
\usage{
params_nn_gpu(
  dist_metric = c("euclidean", "cosine"),
  n_list = NULL,
  n_probes = NULL,
  k = NULL,
  k_build = NULL,
  n_tree = NULL,
  delta = 0.001,
  rho = NULL,
  beam_width = NULL,
  max_beam_iters = NULL,
  n_entry_points = NULL
)
}
\arguments{
\item{dist_metric}{Character. The distance metric to use. Defaults to
\code{"euclidean"}.}

\item{n_list}{Optional integer. Number of clusters to use for IVF. If \code{NULL},
will default to \code{sqrt(n)}.}

\item{n_probes}{Optional integer. Number of clusters to probe for IVF. If
\code{NULL}, will default to \code{sqrt(n_list)}.}

\item{k}{Optional integer. Number of nearest neighbours to search for.}

\item{k_build}{Optional integer. Number of nearest neighbours to use
during graph build.}

\item{n_tree}{Optional integer. Number of trees for graph build.}

\item{delta}{Float. Precision parameter for NN descent. Defaults to \code{0.001}.}

\item{rho}{Optional float. Sample rate parameter for NN descent.}

\item{beam_width}{Optional integer. Beam width for beam search.}

\item{max_beam_iters}{Optional integer. Maximum number of beam search
iterations.}

\item{n_entry_points}{Optional integer. Number of entry points for beam
search.}
}
\value{
A list with the GPU nearest neighbour parameters.
}
\description{
Wrapper function to generate GPU nearest neighbour parameters
}
