% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_cor_gpu}
\alias{rs_cor_gpu}
\title{GPU-accelerated correlation calculations}
\usage{
rs_cor_gpu(x, spearman, verbose)
}
\arguments{
\item{x}{Numerical matrix. The matrix for which to calculate the column
pairwise correlation matrix.}

\item{spearman}{Boolean. Shall the Spearman correlation be calculated
instead of Pearson.}

\item{verbose}{Boolean. Controls verbosity of the function.}
}
\value{
The correlation matrix
}
\description{
GPU-accelerated correlation calculations
}
