% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_cov_gpu}
\alias{rs_cov_gpu}
\title{GPU-accelerated covariance calculations}
\usage{
rs_cov_gpu(x, verbose)
}
\arguments{
\item{x}{Numerical matrix. The matrix for which to calculate the column
pairwise covariance matrix.}

\item{verbose}{Boolean. Controls verbosity of the function.}
}
\value{
The covariance matrix
}
\description{
GPU-accelerated covariance calculations
}
