% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_deserialise_parametric_umap}
\alias{rs_deserialise_parametric_umap}
\title{Deserialises raw bytes to a trained UMAP model.}
\usage{
rs_deserialise_parametric_umap(bytes)
}
\arguments{
\item{bytes}{The raw byte sequence}
}
\value{
Returns
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Deserialises a trained parametric UMAP model from points and returns an R
object.
}
\keyword{internal}
