% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_harmony_v2_gpu}
\alias{rs_harmony_v2_gpu}
\title{Harmony batch correction in Rust (version 2, GPU-accelerated)}
\usage{
rs_harmony_v2_gpu(pca, harmony_params, batch_labels, seed, verbose)
}
\arguments{
\item{pca}{Numerical matrix, i.e., the PCA matrix you want to correct.}

\item{harmony_params}{List. The parameters for the Harmony (v2) GPU algorithm.}

\item{batch_labels}{List. Must contain exactly one element: a 0-indexed
integer vector representing the batch effects you wish to regress out.}

\item{seed}{Integer. Seed for reproducibility purposes.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
The batch-corrected Harmony (v2) embedding space.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function implements the GPU-accelerated version 2 Harmony algorithm
from Patikas, et al., 2026. Only a single batch covariate is supported.
}
