% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sc_pca_sparse_gpu}
\alias{rs_sc_pca_sparse_gpu}
\title{Calculates sparse PCA for single cell}
\usage{
rs_sc_pca_sparse_gpu(
  f_path_gene,
  f_path_cell,
  no_pcs,
  pca_params,
  cell_indices,
  gene_indices,
  seed,
  verbose
)
}
\arguments{
\item{f_path_gene}{String. Path to the \code{counts_genes.bin} file.}

\item{f_path_cell}{String. Path to the \code{counts_cells.bin} file. Used if
you wish to use the PFlogPF transformation.}

\item{no_pcs}{Integer. Number of PCs to calculate.}

\item{pca_params}{Named list. Contains the parameters to use for this PCA
run. (Randomised will ignore, as gpu only supports randomised.)}

\item{cell_indices}{Integer. The cell indices to use. (0-indexed!)}

\item{gene_indices}{Integer. The gene indices to use. (0-indexed!)}

\item{seed}{Integer. Random seed for the randomised SVD.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A list with with the following items
\itemize{
\item scores - The samples projected on the PCA space (solved via sparse
SVD).
\item loadings - The loadings of the features for the PCA (solved via
sparse SVD).
\item singular_values - The singular values for the PCA (solved via sparse
SVD).
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Helper function that will calculate sparse PCA without scaling the data.
This has the advantage that you avoid creating a large dense matrix due
to scaling; however, it has the disadvantage that the first PC will be
heavily influenced by average expression. If random_svd is set to \code{FALSE},
Lanczos iterations will be used to solve the SVD; if random_svd is set
to \code{TRUE}, the randomised version will be used with multiplication of the
initial sparse matrix with a much smaller random dense matrix, avoiding
holding a large dense matrix in memory.
}
\keyword{internal}
