% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_tsne_from_knn_gpu}
\alias{rs_tsne_from_knn_gpu}
\title{tSNE implementation from a pre-computed kNN graph}
\usage{
rs_tsne_from_knn_gpu(
  embd,
  knn_data,
  n_dim,
  perplexity,
  approx_type,
  tsne_params,
  seed,
  use_high_precision,
  verbose
)
}
\arguments{
\item{embd}{Numerical matrix. The data to use to generate the embeddings.
Should be of dimensions samples x features.}

\item{knn_data}{\code{NearestNeighbours} class from R.}

\item{n_dim}{Integer. Number of tSNE dimensions to return. Needs to be two,
others are not supported.}

\item{perplexity}{Numeric. The tSNE perplexity parameter.}

\item{approx_type}{String. One of \code{c("fft", "bh")}. Which of the two
approximations to use.}

\item{tsne_params}{Named list. List that contains all of the key parameters
for the tSNE generation.}

\item{seed}{Integer. Seed for reproducibility.}

\item{use_high_precision}{Optional logical. Controls \code{fp32} vs \code{fp64} for.
If \code{NULL} will use sensible default thresholding.}

\item{verbose}{Integer. If \code{0L} -> silent or \code{1L} for normal verbosity; \code{2L}
for detailed verbosity.}
}
\value{
The tSNE embeddings.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Wraps the tSNE implementation in manifolds-rs. You have two optimiser
options: \code{"bh"}, which tends to be faster on smaller data sets, and \code{"fft"}
for large data sets. This version takes a pre-computed kNN graph, please see
\code{\link[manifoldsR:new_nearest_neighbour]{new_nearest_neighbour()}}.
}
