% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_gpu.R
\name{umap_gpu}
\alias{umap_gpu}
\title{Rust-based UMAP (GPU)}
\usage{
umap_gpu(
  data,
  knn = NULL,
  n_dim = 2L,
  k = 15L,
  min_dist = 0.5,
  spread = 1,
  knn_method = c("nndescent", "exhaustive", "ivf"),
  nn_params = params_nn_gpu(),
  umap_params = params_umap_gpu(),
  seed = 42L,
  use_high_precision = NULL,
  .verbose = TRUE
)
}
\arguments{
\item{data}{Numerical matrix or data frame. The data to embed of shape
samples x features. Will be coerced to a matrix.}

\item{knn}{Optional \code{NearestNeighbours} class. If provided, UMAP will skip
the k-nearest neighbour graph generation and use this one. Defaults to
\code{NULL}. See \code{\link[manifoldsR:new_nearest_neighbour]{manifoldsR::new_nearest_neighbour()}} for details.}

\item{n_dim}{Integer. Number of dimensions in the embedding space.
Defaults to \code{2L}.}

\item{k}{Integer. Number of nearest neighbours to consider for manifold
approximation. Larger values result in more global structure being
preserved. Defaults to \code{15L}.}

\item{min_dist}{Numeric. Minimum distance between points in the embedding.
Controls how tightly points are packed. Smaller values result in more
clustered embeddings. Must be >= 0. Defaults to \code{0.5}. If you use SGD,
consider reducing this!}

\item{spread}{Numeric. Effective scale of embedded points. Determines the
scale at which embedded points will be spread out. Defaults to \code{1.0}.}

\item{knn_method}{Character. (Approximate) Nearest neighbour method to use.
One of \code{"exhaustive"}, \code{"ivf"} or \code{"nndescent_gpu"}.}

\item{umap_params}{Named list. UMAP (GPU) algorithm parameters, see
\code{\link[=params_umap_gpu]{params_umap_gpu()}}.}

\item{seed}{Integer. Random seed for reproducibility. Defaults to \code{42L}.}

\item{use_high_precision}{Optional boolean. Gives fine-grained control over
\code{fp32} vs \code{fp64} usage.}

\item{.verbose}{Logical. Controls verbosity. Defaults to \code{TRUE}.}

\item{nn_params_gpu}{Named list. Nearest neighbour search parameters, see
\code{\link[=params_nn_gpu]{params_nn_gpu()}}.}
}
\value{
A numerical matrix with dimensions samples x n_dim containing
the UMAP embedding.
}
\description{
Performs UMAP dimensionality reduction on the input data.
This function provides a user-friendly interface with input validation
before calling the Rust implementation. Leverages GPU-accelerated kNN
searches and in the default setting also uses a GPU-accelerated Adam
optimiser for the embedding.
}
